AGENT-97: Fix mapping type assumed by Mypy
